# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcPredict Prediction Market API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MarketStatus = (typeof MarketStatus)[keyof typeof MarketStatus];

export const MarketStatus = {
  open: "open",
  closed: "closed",
  resolved: "resolved",
} as const;

export type MarketOutcome =
  | (typeof MarketOutcome)[keyof typeof MarketOutcome]
  | null;

export const MarketOutcome = {
  yes: "yes",
  no: "no",
} as const;

export interface Market {
  id: number;
  title: string;
  description: string;
  category: string;
  status: MarketStatus;
  outcome?: MarketOutcome;
  /** Current YES probability (0-1) */
  yesPrice: number;
  /** Current NO probability (0-1) */
  noPrice: number;
  totalVolume: number;
  betCount?: number;
  createdAt: string;
  closesAt: string;
  creatorId: number;
}

export type BetPosition = (typeof BetPosition)[keyof typeof BetPosition];

export const BetPosition = {
  yes: "yes",
  no: "no",
} as const;

export type BetStatus = (typeof BetStatus)[keyof typeof BetStatus];

export const BetStatus = {
  pending: "pending",
  won: "won",
  lost: "lost",
} as const;

export interface Bet {
  id: number;
  marketId: number;
  userId: number;
  username?: string;
  position: BetPosition;
  amount: number;
  price: number;
  payout: number | null;
  status: BetStatus;
  createdAt: string;
}

export interface User {
  id: number;
  username: string;
  avatarUrl?: string | null;
  balance: number;
  totalBets: number;
  totalWins: number;
  createdAt: string;
}

export type MarketDetail = Market & {
  recentBets?: Bet[];
  creator?: User;
};

export interface MarketList {
  markets: Market[];
  total: number;
}

export interface CreateMarketBody {
  title: string;
  description: string;
  category: string;
  closesAt: string;
  creatorId: number;
}

export type ResolveMarketBodyOutcome =
  (typeof ResolveMarketBodyOutcome)[keyof typeof ResolveMarketBodyOutcome];

export const ResolveMarketBodyOutcome = {
  yes: "yes",
  no: "no",
} as const;

export interface ResolveMarketBody {
  outcome: ResolveMarketBodyOutcome;
}

export type PlaceBetBodyPosition =
  (typeof PlaceBetBodyPosition)[keyof typeof PlaceBetBodyPosition];

export const PlaceBetBodyPosition = {
  yes: "yes",
  no: "no",
} as const;

export interface PlaceBetBody {
  userId: number;
  position: PlaceBetBodyPosition;
  amount: number;
}

export interface CreateUserBody {
  username: string;
  avatarUrl?: string;
}

export interface UserPortfolio {
  user: User;
  totalInvested: number;
  totalPayout: number;
  netPnl: number;
  winRate: number;
  positions: Bet[];
}

export type PlatformStatsCategoryBreakdownItem = {
  category: string;
  count: number;
  volume: number;
};

export interface PlatformStats {
  totalMarkets: number;
  openMarkets: number;
  totalBets: number;
  totalVolume: number;
  totalUsers: number;
  categoryBreakdown: PlatformStatsCategoryBreakdownItem[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  avatarUrl?: string | null;
  netPnl: number;
  totalBets: number;
  winRate: number;
  balance?: number;
}

export type ListMarketsParams = {
  status?: ListMarketsStatus;
  category?: string;
  limit?: number;
  offset?: number;
};

export type ListMarketsStatus =
  (typeof ListMarketsStatus)[keyof typeof ListMarketsStatus];

export const ListMarketsStatus = {
  open: "open",
  closed: "closed",
  resolved: "resolved",
} as const;